pyproject.toml
[project]
name = "spice_rl"
version = "0.1.0"
description = "An OpenEnv RL environment for DC-DC converter design using SPICE simulation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}
dependencies = [
    "openenv-core>=0.1.0",
    "spicelib>=1.2.0",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "pydantic>=2.0",
    "openai>=1.0",
    "numpy>=1.24",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["server*", "spice_rl*"]
